+++ src/main/java/com/terminalvelocitycabbage/engine/client/renderer/shapes/TexturedModelCuboid.java
@@ -20,7 +20,7 @@ public class TexturedModelCuboid {
 				topTL, topBL, topBR, topTR,
 				bottomTL, bottomBL, bottomBR, bottomTR,
 			},
-			new short[]{
+			new int[]{
 				0, 1, 2, 2, 3, 0,
 				4, 5, 6, 6, 7, 4,
 				8, 9, 10, 10, 11, 8,
